requirements.txt
asgiref==3.8.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
Pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
dj_database_url==2.3.0
django-crontab==0.7.1
django-modeltranslation==0.19.12
django-rosetta==0.10.1
setuptools>=58.0.0
drf-yasg
gunicorn==21.2.0